move all log functions to separate file for clarify
